DistributorController rewritten to use DTO

+++ src/main/java/com/moviesdb/moviesdb/DTOs/dto/TVShowDTO.java
@@ -22,7 +22,7 @@ public class TVShowDTO extends WatchableBaseEntityDTO {
 
     private Set<Long> distributors;
 
-    public TVShowDTO(Long id, String name, String description, Short ageRestriction, Integer duration, Byte rating, Set<Long> actors, Integer totalEpisodes, Integer totalSeasons, Long director_id, Set<Long> distributors) {
+    public TVShowDTO(Long id, String name, String description, Short ageRestriction, Short duration, Byte rating, Set<Long> actors, Integer totalEpisodes, Integer totalSeasons, Long director_id, Set<Long> distributors) {
         super(id, name, description, ageRestriction, duration, rating);
         this.actors = actors;
         this.totalEpisodes = totalEpisodes;
